models/proto-gcn/requirements.txt
decord>=0.6.0
fvcore
matplotlib
mmcv-full==1.5.0
mmdet==2.23.0
mmpose==0.24.0
moviepy
numpy>=1.19.5
opencv-contrib-python
opencv-python
importlib_metadata
pymemcache
pyarrow>=10.0.0
scipy
torch>=1.5
tqdm
yapf==0.32.0